Clean up GameMenu and move things to LocalMatchSelect
