+++ _main_project/static/js/views/GameMenu.js
@@ -1,13 +1,11 @@
 import AbstractView from "./AbstractView.js";
 import { navigateTo } from "../helpers/helpers.js";
-import { gameBoard } from "../index.js";
 
 export default class extends AbstractView {
   constructor(params) {
     super(params);
     this.setTitle("Game Menu");
     this.name = "GameMenu";
-    this.container = null;
   }
 
   getDomElements() {
@@ -15,8 +13,8 @@ export default class extends AbstractView {
     document.getElementById("gameModal").style.display = "none";
 
     // Create a this.container div
-    this.container = document.createElement('div');
-    this.container.classList.add('text-container', 'game-select-button-div');
+    const container = document.createElement('div');
+    container.classList.add('text-container', 'game-select-button-div');
 
     const setTitle = document.createElement('h1');
     setTitle.textContent = 'Select a game mode:';
@@ -43,23 +41,19 @@ export default class extends AbstractView {
     tournament_button.type = 'select';
     tournament_button.textContent = 'Tournament';
 
-    // Append the paragraph to the this.container
-    this.container.appendChild(setTitle);
-    this.container.appendChild(local_match_button);
-    this.container.appendChild(multi_player_button);
-    this.container.appendChild(tournament_button);
+    // Append the paragraph to the container
+    container.appendChild(setTitle);
+    container.appendChild(local_match_button);
+    container.appendChild(multi_player_button);
+    container.appendChild(tournament_button);
 
-    return this.container;
+    return container;
   }
 
   async afterRender() {
 
     document.getElementById('localMatchBtn').addEventListener('click', async() => {
 			navigateTo('/local_match_select/');
-      // this.chooseMode((mode) => {
-      //   gameBoard.startSinglePlayerGame(mode);
-      //   gameModal.style.display = 'flex';
-      // });
 		});
 
 		document.getElementById('multiPlayerBtn').addEventListener('click', async() => {
@@ -67,50 +61,7 @@ export default class extends AbstractView {
 		});
 
     document.getElementById('tournamentBtn').addEventListener('click', () => {
-      // load in Remote game on the backend.
-      console.log('Join Tournament');
       navigateTo('/tournament_select/');
-      // navigateTo('/tournament_setup_join/');
     });
   }
-
-	chooseMode(callback) {
-		// check if the mode dialog is already open
-		if (document.querySelector('.mode-dialog')) return;
-
-		// Hiding all elements in the container
-		this.container.querySelectorAll('*').forEach(element => {
-			element.style.display = 'none';
-		});
-
-		const modeDialog = document.createElement('div');
-		modeDialog.classList.add('mode-dialog');
-	
-		const modeText = document.createElement('p');
-		modeText.textContent = 'Choose Game Mode for Single Player:';
-		modeDialog.appendChild(modeText);
-	
-		// const modes = ['Single with 2 paddles on one keyboard', 'Aginst AI'];
-		const modes = {
-			'Single with control over 2 paddles on one keyboard': 'single',
-			'Play Aginst AI': 'ai'
-		};
-
-		// modes.forEach(mode => {
-		Object.keys(modes).forEach(displayText => {
-			const modeButton = document.createElement('button');
-			modeButton.textContent = displayText;
-			modeButton.classList.add('mode-button', "game-select-button");
-			modeButton.addEventListener('click', () => {
-				this.container.classList.remove('hidden-elements');
-				this.container.removeChild(modeDialog);
-				callback(modes[displayText]);
-			});
-			modeDialog.appendChild(modeButton);
-		});
-	
-		this.container.appendChild(modeDialog);
-		this.container.classList.add('hidden-elements');
-	}
-
 }
